+++ client/acp/src/model/setting-fields.js
@@ -3,3 +3,5 @@ export const ACTIVITY_LIMIT = 'activityLimit';
 export const MAX_REWARDS_PER_ACCOUNT = 'maxRewardsPerAccount';
 
 export const MAX_REWARDS_PER_POST = 'maxRewardsPerPost';
+
+export const NOTIFICATION_TEXT = 'notificationText';